测试消息事件；
