+++ project-mongo/src/main/java/cn/taike/mongo/event/msg/Message.java
@@ -1,10 +1,12 @@
-package cn.taike.mongo.event;
+package cn.taike.mongo.event.msg;
 
+import lombok.Data;
 import org.springframework.context.ApplicationEvent;
 
 /**
  * Created by huayandong on 17/8/22.
  */
+@Data
 public class Message extends ApplicationEvent {
 
     private String name;
@@ -13,11 +15,4 @@ public class Message extends ApplicationEvent {
         super(source);
     }
 
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
 }
